fixing > between list and int error
I was getting an error stating that the > symbol could not be present between a list and an integer. This seems to have fixed it, but I'm not sure if that is just covering a different problem in the line where we assign the value to the num_faces_answered variable.