Need FaceSource to automatically send out first FacePair
I think with the changes I made we will now need the FaceSource to send out the first set of FacePairs on its own automatically when the participant arrives to the Face Page. Though I'm not sure how to do this, or where the instructions for the Questionnaire soure to do so are.