Changing wait_div
I'm not sure how to make different wait divs for the various stages of round 1.
I would like to have an instructions page before they begin choosing faces, and a different wait div between when they are choosing faces for the first time vs. seeing the faces for a second time